--- src/main/java/executor/service/model/StepTypes.java
@@ -1,17 +1,0 @@
-package executor.service.model;
-
-public enum StepTypes {
-  CLICK_CSS("clickCss"),
-  SLEEP("sleep"),
-  CLICK_XPATH("clickXpath");
-
-  private final String name;
-
-  StepTypes(String name) {
-    this.name = name;
-  }
-
-  public String getName() {
-    return name;
-  }
-}
