+++ src/main/java/executor/service/service/impl/StepExecutionClickXpathImpl.java
@@ -1,7 +1,7 @@
 package executor.service.service.impl;
 
 import executor.service.model.Step;
-import executor.service.model.StepTypes;
+import executor.service.model.StepAction;
 import executor.service.service.StepExecutionClickXpath;
 import org.openqa.selenium.WebDriver;
 
@@ -9,7 +9,7 @@ public class StepExecutionClickXpathImpl implements StepExecutionClickXpath {
 
   @Override
   public String getStepAction() {
-    return StepTypes.CLICK_XPATH.getName();
+    return StepAction.CLICK_XPATH.getName();
   }
 
   @Override
